Upstream version 0.8.1

# gpg: Signature made Mit 27 Jan 2016 18:15:03 CET
# gpg:                using RSA key 0x14225108C1D2569F
# gpg: Good signature from "Hendrik Jaeger <henk@hnjs.ch>" [ultimate]
# gpg:                 aka "Hendrik Jaeger <hjaeger@hnjs.ch>" [ultimate]
# gpg:                 aka "Hendrik Jaeger <henk@netwichtig.de>" [ultimate]

+++ src/io/iop.c
@@ -16,66 +16,84 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA
  */
 
-#include "iop.h"
-
 #ifdef _WIN32
 #define REQUIRED_WINVER 0x0600
 #include "../utils/windefs.h"
 #include <windows.h>
 #endif
 
-#include <sys/stat.h> /* stat chmod() mkdir() */
+#include "iop.h"
+
+#ifndef _WIN32
+#include <sys/ioctl.h> /* ioctl() */
+#endif
+#include <sys/stat.h> /* stat */
 #include <sys/types.h> /* mode_t */
 #include <unistd.h> /* rmdir() symlink() unlink() */
 
-#include <errno.h> /* EEXIST errno */
+#include <assert.h> /* assert() */
+#include <errno.h> /* EEXIST ENOENT EISDIR errno */
 #include <stddef.h> /* NULL size_t */
-#include <stdio.h> /* FILE fpos_t fclose() fgetpos() fopen() fread() fseek()
-                      fsetpos() fwrite() rename() snprintf() */
+#include <stdio.h> /* FILE fpos_t fclose() fgetpos() fread() fseek() fsetpos()
+                      fwrite() snprintf() */
 #include <stdlib.h> /* free() */
-#include <string.h> /* strchr() */
+#include <string.h> /* strchr() strerror() */
 
+#include "../compat/fs_limits.h"
+#include "../compat/os.h"
+#include "../ui/cancellation.h"
 #include "../utils/fs.h"
-#include "../utils/fs_limits.h"
 #include "../utils/log.h"
 #include "../utils/macros.h"
 #include "../utils/path.h"
 #include "../utils/str.h"
+#include "../utils/utf8.h"
+#include "../utils/utils.h"
 #include "../background.h"
-#include "../ui.h"
+#include "private/ioe.h"
 #include "private/ioeta.h"
 #include "ioc.h"
 
 /* Amount of data to transfer at once. */
 #define BLOCK_SIZE 32*1024
 
+static int clone_file(int dest_fd, int src_fd);
 #ifdef _WIN32
 static DWORD CALLBACK win_progress_cb(LARGE_INTEGER total,
 		LARGE_INTEGER transferred, LARGE_INTEGER stream_size,
 		LARGE_INTEGER stream_transfered, DWORD stream_num, DWORD reason,
 		HANDLE src_file, HANDLE dst_file, LPVOID param);
 #endif
+static IoErrCbResult sig_err(io_args_t *const args, int *result,
+		const char path[], int error_code, const char msg[]);
 
 int
 iop_mkfile(io_args_t *const args)
 {
+	FILE *f;
 	const char *const path = args->arg1.path;
 
-#ifndef _WIN32
-	FILE *const f = fopen(path, "wb");
-	return (f == NULL) ? -1 : fclose(f);
-#else
-	HANDLE file;
-
-	file = CreateFileA(path, 0, 0, NULL, CREATE_NEW, FILE_ATTRIBUTE_NORMAL, NULL);
-	if(file == INVALID_HANDLE_VALUE)
+	if(path_exists(path, DEREF))
 	{
+		(void)ioe_errlst_append(&args->result.errors, path, EEXIST,
+				strerror(EEXIST));
 		return -1;
 	}
 
-	CloseHandle(file);
+	f = os_fopen(path, "wb");
+	if(f == NULL)
+	{
+		(void)ioe_errlst_append(&args->result.errors, path, errno, strerror(errno));
+		return -1;
+	}
+
+	if(fclose(f) != 0)
+	{
+		(void)ioe_errlst_append(&args->result.errors, path, errno, strerror(errno));
+		return -1;
+	}
+
 	return 0;
-#endif
 }
 
 int
@@ -103,25 +121,37 @@ iop_mkdir(io_args_t *const args)
 				continue;
 			}
 
-			/* Create intermediate directories with 0755 permissions. */
-			if(make_dir(partial_path, 0755) != 0)
+			/* Create intermediate directories with 0700 permissions. */
+			if(os_mkdir(partial_path, S_IRWXU) != 0)
 			{
+				(void)ioe_errlst_append(&args->result.errors, partial_path, errno,
+						strerror(errno));
+
 				free(partial_path);
 				return -1;
 			}
 		}
 
-		free(partial_path);
 #ifndef _WIN32
-		return chmod(path, mode);
-#else
-		return 0;
+		if(os_chmod(path, mode) != 0)
+		{
+			(void)ioe_errlst_append(&args->result.errors, partial_path, errno,
+					strerror(errno));
+			free(partial_path);
+			return 1;
+		}
 #endif
+
+		free(partial_path);
+		return 0;
 	}
-	else
+
+	if(os_mkdir(path, mode) != 0)
 	{
-		return make_dir(path, mode);
+		(void)ioe_errlst_append(&args->result.errors, path, errno, strerror(errno));
+		return 1;
 	}
+	return 0;
 }
 
 int
@@ -132,24 +162,57 @@ iop_rmfile(io_args_t *const args)
 	uint64_t size;
 	int result;
 
-	ioeta_update(args->estim, path, 0, 0);
+	ioeta_update(args->estim, path, path, 0, 0);
 
 	size = get_file_size(path);
 
 #ifndef _WIN32
-	result = unlink(path);
+	do
+	{
+		result = unlink(path);
+		if(result != 0)
+		{
+			if(sig_err(args, &result, path, errno, strerror(errno)) == IO_ECR_RETRY)
+			{
+				continue;
+			}
+		}
+
+		break;
+	}
+	while(1);
 #else
 	{
-		const DWORD attributes = GetFileAttributesA(path);
-		if(attributes & FILE_ATTRIBUTE_READONLY)
+		wchar_t *const utf16_path = utf8_to_utf16(path);
+		const DWORD attributes = GetFileAttributesW(utf16_path);
+
+		do
 		{
-			SetFileAttributesA(path, attributes & ~FILE_ATTRIBUTE_READONLY);
+			if(attributes & FILE_ATTRIBUTE_READONLY)
+			{
+				SetFileAttributesW(utf16_path, attributes & ~FILE_ATTRIBUTE_READONLY);
+			}
+			result = (DeleteFileW(utf16_path) == FALSE);
+
+			if(result != 0)
+			{
+				/* FIXME: use real system error message here. */
+				if(sig_err(args, &result, path, IO_ERR_UNKNOWN,
+							"File removal failed") == IO_ECR_RETRY)
+				{
+					continue;
+				}
+			}
+
+			break;
 		}
+		while(1);
+
+		free(utf16_path);
 	}
-	result = !DeleteFile(path);
 #endif
 
-	ioeta_update(args->estim, path, 1, size);
+	ioeta_update(args->estim, NULL, NULL, 1, size);
 
 	return result;
 }
@@ -161,15 +224,49 @@ iop_rmdir(io_args_t *const args)
 
 	int result;
 
-	ioeta_update(args->estim, path, 0, 0);
+	ioeta_update(args->estim, path, path, 0, 0);
 
 #ifndef _WIN32
-	result = rmdir(path);
+	do
+	{
+		result = rmdir(path);
+		if(result != 0)
+		{
+			if(sig_err(args, &result, path, errno, strerror(errno)) == IO_ECR_RETRY)
+			{
+				continue;
+			}
+		}
+
+		break;
+	}
+	while(1);
 #else
-	result = RemoveDirectory(path) == 0;
+	{
+		wchar_t *const utf16_path = utf8_to_utf16(path);
+
+		do
+		{
+			result = (RemoveDirectoryW(utf16_path) == FALSE);
+			if(result != 0)
+			{
+				/* FIXME: use real system error message here. */
+				if(sig_err(args, &result, path, IO_ERR_UNKNOWN,
+							"Directory removal failed") == IO_ECR_RETRY)
+				{
+					continue;
+				}
+			}
+
+			break;
+		}
+		while(1);
+
+		free(utf16_path);
+	}
 #endif
 
-	ioeta_update(args->estim, path, 1, 0);
+	ioeta_update(args->estim, NULL, NULL, 1, 0);
 
 	return result;
 }
@@ -180,32 +277,63 @@ iop_cp(io_args_t *const args)
 	const char *const src = args->arg1.src;
 	const char *const dst = args->arg2.dst;
 	const IoCrs crs = args->arg3.crs;
+	const io_confirm confirm = args->confirm;
 	const int cancellable = args->cancellable;
+	struct stat st;
 
 	char block[BLOCK_SIZE];
 	FILE *in, *out;
 	size_t nread;
 	int error;
+	int cloned;
 	struct stat src_st;
 	const char *open_mode = "wb";
 
-	ioeta_update(args->estim, src, 0, 0);
+	ioeta_update(args->estim, src, dst, 0, 0);
 
 #ifdef _WIN32
 	if(is_symlink(src) || crs != IO_CRS_APPEND_TO_FILES)
 	{
 		DWORD flags;
 		int error;
+		wchar_t *utf16_src, *utf16_dst;
 
 		flags = COPY_FILE_COPY_SYMLINK;
 		if(crs == IO_CRS_FAIL)
 		{
 			flags |= COPY_FILE_FAIL_IF_EXISTS;
 		}
+		else if(path_exists(dst, DEREF))
+		{
+			/* Ask user whether to overwrite destination file. */
+			if(confirm != NULL && !confirm(args, src, dst))
+			{
+				return 0;
+			}
+		}
 
-		error = CopyFileExA(src, dst, &win_progress_cb, args, NULL, flags) == 0;
+		utf16_src = utf8_to_utf16(src);
+		utf16_dst = utf8_to_utf16(dst);
 
-		ioeta_update(args->estim, src, 1, 0);
+		error = CopyFileExW(utf16_src, utf16_dst, &win_progress_cb, args, NULL,
+				flags) == 0;
+
+		if(error)
+		{
+			/* FIXME: use real system error message here. */
+			(void)ioe_errlst_append(&args->result.errors, dst, IO_ERR_UNKNOWN,
+					"Copy file failed");
+		}
+
+		free(utf16_src);
+		free(utf16_dst);
+
+		if(error == 0)
+		{
+			clone_timestamps(dst, src, NULL);
+		}
+
+		ioeta_update(args->estim, NULL, NULL, 1, 0);
 
 		return error;
 	}
@@ -216,33 +344,68 @@ iop_cp(io_args_t *const args)
 	if(is_symlink(src))
 	{
 		char link_target[PATH_MAX];
+		int error;
 
-		io_args_t args =
-		{
+		io_args_t ln_args = {
 			.arg1.path = link_target,
 			.arg2.target = dst,
 			.arg3.crs = crs,
 
 			.cancellable = cancellable,
+
+			.result = args->result,
 		};
 
 		if(get_link_target(src, link_target, sizeof(link_target)) != 0)
 		{
+			(void)ioe_errlst_append(&args->result.errors, src, IO_ERR_UNKNOWN,
+					"Failed to get symbolic link target");
 			return 1;
 		}
 
-		return iop_ln(&args);
+		error = iop_ln(&ln_args);
+		args->result = ln_args.result;
+
+		if(error != 0)
+		{
+			(void)ioe_errlst_append(&args->result.errors, src, IO_ERR_UNKNOWN,
+					"Failed to make symbolic link");
+			return 1;
+		}
+		return 0;
 	}
 
 	if(is_dir(src))
 	{
+		(void)ioe_errlst_append(&args->result.errors, src, EISDIR,
+				strerror(EISDIR));
 		return 1;
 	}
 
-	in = fopen(src, "rb");
-	if(in == NULL)
+	if(os_stat(src, &st) != 0)
 	{
+		(void)ioe_errlst_append(&args->result.errors, src, errno, strerror(errno));
 		return 1;
+	}
+
+#ifndef _WIN32
+	/* Fifo/socket/device files don't need to be opened, their content is not
+	 * accessed. */
+	if(S_ISFIFO(st.st_mode) || S_ISSOCK(st.st_mode) || S_ISBLK(st.st_mode) ||
+			S_ISCHR(st.st_mode))
+	{
+		in = NULL;
+	}
+	else
+#endif
+	{
+		in = os_fopen(src, "rb");
+		if(in == NULL)
+		{
+			(void)ioe_errlst_append(&args->result.errors, src, errno,
+					strerror(errno));
+			return 1;
+		}
 	}
 
 	if(crs == IO_CRS_APPEND_TO_FILES)
@@ -251,10 +414,32 @@ iop_cp(io_args_t *const args)
 	}
 	else if(crs != IO_CRS_FAIL)
 	{
-		const int ec = unlink(dst);
+		int ec;
+
+		if(path_exists(dst, DEREF))
+		{
+			/* Ask user whether to overwrite destination file. */
+			if(confirm != NULL && !confirm(args, src, dst))
+			{
+				if(in != NULL && fclose(in) != 0)
+				{
+					(void)ioe_errlst_append(&args->result.errors, src, errno,
+							strerror(errno));
+				}
+				return 0;
+			}
+		}
+
+		ec = unlink(dst);
 		if(ec != 0 && errno != ENOENT)
 		{
-			fclose(in);
+			(void)ioe_errlst_append(&args->result.errors, dst, errno,
+					strerror(errno));
+			if(in != NULL && fclose(in) != 0)
+			{
+				(void)ioe_errlst_append(&args->result.errors, src, errno,
+						strerror(errno));
+			}
 			return ec;
 		}
 
@@ -263,35 +448,83 @@ iop_cp(io_args_t *const args)
 		 * but this approach has disadvantage of requiring more free space on
 		 * destination file system. */
 	}
-	else if(path_exists(dst))
+	else if(path_exists(dst, DEREF))
 	{
-		fclose(in);
+		(void)ioe_errlst_append(&args->result.errors, src, EEXIST,
+				strerror(EEXIST));
+		if(in != NULL && fclose(in) != 0)
+		{
+			(void)ioe_errlst_append(&args->result.errors, src, errno,
+					strerror(errno));
+		}
 		return 1;
 	}
 
-	out = fopen(dst, open_mode);
+#ifndef _WIN32
+	/* Replicate fifo without even opening it. */
+	if(S_ISFIFO(st.st_mode))
+	{
+		if(mkfifo(dst, st.st_mode & 07777) != 0)
+		{
+			(void)ioe_errlst_append(&args->result.errors, src, errno,
+					strerror(errno));
+			return 1;
+		}
+		return 0;
+	}
+
+	/* Replicate socket or device file without even opening it. */
+	if(S_ISSOCK(st.st_mode) || S_ISBLK(st.st_mode) || S_ISCHR(st.st_mode))
+	{
+		if(mknod(dst, st.st_mode & (S_IFMT | 07777), st.st_rdev) != 0)
+		{
+			(void)ioe_errlst_append(&args->result.errors, src, errno,
+					strerror(errno));
+			return 1;
+		}
+		return 0;
+	}
+#endif
+
+	out = os_fopen(dst, open_mode);
 	if(out == NULL)
 	{
-		fclose(in);
+		(void)ioe_errlst_append(&args->result.errors, dst, errno, strerror(errno));
+		if(fclose(in) != 0)
+		{
+			(void)ioe_errlst_append(&args->result.errors, src, errno,
+					strerror(errno));
+		}
 		return 1;
 	}
 
 	error = 0;
+	cloned = 0;
 
 	if(crs == IO_CRS_APPEND_TO_FILES)
 	{
 		fpos_t pos;
+		/* The following line is required for stupid Windows sometimes.  Why?
+		 * Probably because it's stupid...  Won't harm other systems. */
+		fseek(out, 0, SEEK_END);
 		error = fgetpos(out, &pos) != 0 || fsetpos(in, &pos) != 0;
 
 		if(!error)
 		{
-			ioeta_update(args->estim, src, 0, get_file_size(dst));
+			ioeta_update(args->estim, NULL, NULL, 0, get_file_size(dst));
+		}
+	}
+	else if(args->arg4.fast_file_cloning)
+	{
+		if(clone_file(fileno(out), fileno(in)) == 0)
+		{
+			cloned = 1;
 		}
 	}
 
 	/* TODO: use sendfile() if platform supports it. */
 
-	while((nread = fread(&block, 1, sizeof(block), in)) != 0U)
+	while(!cloned && (nread = fread(&block, 1, sizeof(block), in)) != 0U)
 	{
 		if(cancellable && ui_cancellation_requested())
 		{
@@ -301,24 +534,64 @@ iop_cp(io_args_t *const args)
 
 		if(fwrite(&block, 1, nread, out) != nread)
 		{
+			(void)ioe_errlst_append(&args->result.errors, dst, errno,
+					strerror(errno));
 			error = 1;
 			break;
 		}
 
-		ioeta_update(args->estim, src, 0, nread);
+		ioeta_update(args->estim, NULL, NULL, 0, nread);
 	}
-
-	fclose(in);
-	fclose(out);
-
-	if(error == 0 && lstat(src, &src_st) == 0)
+	if(!cloned && nread == 0U && !feof(in) && ferror(in))
 	{
-		error = chmod(dst, src_st.st_mode & 07777);
+		(void)ioe_errlst_append(&args->result.errors, src, errno, strerror(errno));
 	}
 
-	ioeta_update(args->estim, src, 1, 0);
+	if(fclose(in) != 0)
+	{
+		(void)ioe_errlst_append(&args->result.errors, src, errno, strerror(errno));
+	}
+	if(fclose(out) != 0)
+	{
+		(void)ioe_errlst_append(&args->result.errors, dst, errno, strerror(errno));
+	}
+
+	if(error == 0 && os_lstat(src, &src_st) == 0)
+	{
+		error = os_chmod(dst, src_st.st_mode & 07777);
+		if(error != 0)
+		{
+			(void)ioe_errlst_append(&args->result.errors, dst, errno,
+					strerror(errno));
+		}
+	}
+
+	if(error == 0)
+	{
+		clone_timestamps(dst, src, &st);
+	}
+
+	ioeta_update(args->estim, NULL, NULL, 1, 0);
 
 	return error;
+}
+
+/* Try to clone file fast on btrfs.  Returns 0 on success, otherwise non-zero is
+ * returned. */
+static int
+clone_file(int dst, int src)
+{
+#ifdef __linux__
+#undef BTRFS_IOCTL_MAGIC
+#define BTRFS_IOCTL_MAGIC 0x94
+#undef BTRFS_IOC_CLONE
+#define BTRFS_IOC_CLONE _IOW(BTRFS_IOCTL_MAGIC, 9, int)
+	return ioctl(dst, BTRFS_IOC_CLONE, src);
+#else
+	(void)dst;
+	(void)src;
+	return -1;
+#endif
 }
 
 #ifdef _WIN32
@@ -333,6 +606,7 @@ static DWORD CALLBACK win_progress_cb(LARGE_INTEGER total,
 	io_args_t *const args = param;
 
 	const char *const src = args->arg1.src;
+	const char *const dst = args->arg2.dst;
 	ioeta_estim_t *const estim = args->estim;
 
 	if(transferred.QuadPart < last_size)
@@ -340,7 +614,7 @@ static DWORD CALLBACK win_progress_cb(LARGE_INTEGER total,
 		last_size = 0;
 	}
 
-	ioeta_update(estim, src, 0, transferred.QuadPart - last_size);
+	ioeta_update(estim, src, dst, 0, transferred.QuadPart - last_size);
 
 	last_size = transferred.QuadPart;
 
@@ -354,10 +628,13 @@ static DWORD CALLBACK win_progress_cb(LARGE_INTEGER total,
 
 #endif
 
+/* TODO: implement iop_chown(). */
 int iop_chown(io_args_t *const args);
 
+/* TODO: implement iop_chgrp(). */
 int iop_chgrp(io_args_t *const args);
 
+/* TODO: implement iop_chmod(). */
 int iop_chmod(io_args_t *const args);
 
 int
@@ -369,12 +646,6 @@ iop_ln(io_args_t *const args)
 
 	int result;
 
-#ifdef _WIN32
-	char cmd[6 + PATH_MAX*2 + 1];
-	char *escaped_path, *escaped_target;
-	char base_dir[PATH_MAX + 2];
-#endif
-
 #ifndef _WIN32
 	result = symlink(path, target);
 	if(result != 0 && errno == EEXIST && overwrite && is_symlink(target))
@@ -383,23 +654,48 @@ iop_ln(io_args_t *const args)
 		if(result == 0)
 		{
 			result = symlink(path, target);
+			if(result != 0)
+			{
+				(void)ioe_errlst_append(&args->result.errors, path, errno,
+						strerror(errno));
+			}
+		}
+		else
+		{
+			(void)ioe_errlst_append(&args->result.errors, target, errno,
+					strerror(errno));
 		}
 	}
-#else
-	if(!overwrite && path_exists(target))
+	else if(result != 0 && errno != 0)
 	{
+		(void)ioe_errlst_append(&args->result.errors, target, errno,
+				strerror(errno));
+	}
+#else
+	char cmd[6 + PATH_MAX*2 + 1];
+	char *escaped_path, *escaped_target;
+	char base_dir[PATH_MAX + 2];
+
+	if(!overwrite && path_exists(target, DEREF))
+	{
+		(void)ioe_errlst_append(&args->result.errors, target, EEXIST,
+				strerror(EEXIST));
 		return -1;
 	}
 
 	if(overwrite && !is_symlink(target))
 	{
+		(void)ioe_errlst_append(&args->result.errors, target, IO_ERR_UNKNOWN,
+				"Target is not a symbolic link");
 		return -1;
 	}
 
-	escaped_path = escape_filename(path, 0);
-	escaped_target = escape_filename(target, 0);
+	escaped_path = shell_like_escape(path, 0);
+	escaped_target = shell_like_escape(target, 0);
 	if(escaped_path == NULL || escaped_target == NULL)
 	{
+		(void)ioe_errlst_append(&args->result.errors, target, IO_ERR_UNKNOWN,
+				"Not enough memory");
 		free(escaped_target);
 		free(escaped_path);
 		return -1;
@@ -407,6 +703,8 @@ iop_ln(io_args_t *const args)
 
 	if(GetModuleFileNameA(NULL, base_dir, ARRAY_LEN(base_dir)) == 0)
 	{
+		(void)ioe_errlst_append(&args->result.errors, target, IO_ERR_UNKNOWN,
+				"Failed to find win_helper");
 		free(escaped_target);
 		free(escaped_path);
 		return -1;
@@ -415,7 +713,13 @@ iop_ln(io_args_t *const args)
 	break_atr(base_dir, '\\');
 	snprintf(cmd, sizeof(cmd), "%s\\win_helper -s %s %s", base_dir, escaped_path,
 			escaped_target);
-	result = system(cmd);
+
+	result = os_system(cmd);
+	if(result != 0)
+	{
+		(void)ioe_errlst_append(&args->result.errors, target, IO_ERR_UNKNOWN,
+				"Running win_helper has failed");
+	}
 
 	free(escaped_target);
 	free(escaped_path);
@@ -424,5 +728,36 @@ iop_ln(io_args_t *const args)
 	return result;
 }
 
+/* Error handler that calls external code to figure out what to do and also logs
+ * the error when needed.  Returns the response. */
+static IoErrCbResult
+sig_err(io_args_t *const args, int *result, const char path[], int error_code,
+		const char msg[])
+{
+	ioerr_cb errors = args->result.errors_cb;
+
+	ioe_err_t err = {
+		.path = (char*)path, .error_code = errno, .msg = (char*)msg,
+	};
+
+	assert(*result != 0 && "The function should be called on error path only.");
+
+	switch((errors == NULL) ? IO_ECR_BREAK : errors(args, &err))
+	{
+		case IO_ECR_RETRY:
+			return IO_ECR_RETRY;
+		case IO_ECR_IGNORE:
+			*result = 0;
+			return IO_ECR_IGNORE;
+		case IO_ECR_BREAK:
+			(void)ioe_errlst_append(&args->result.errors, path, errno, msg);
+			return IO_ECR_BREAK;
+
+		default:
+			assert(0 && "Unknown error handling result.");
+			return IO_ECR_BREAK;
+	}
+}
+
 /* vim: set tabstop=2 softtabstop=2 shiftwidth=2 noexpandtab cinoptions-=(0 : */
-/* vim: set cinoptions+=t0 : */
+/* vim: set cinoptions+=t0 filetype=c : */
